requirements.txt
appdirs==1.4.0
boto==2.42.0
decorator==4.0.11
Flask==0.10
Flask-API-Utils==1.0.2
google-api-python-client==1.6.1
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.9.4
MarkupSafe==0.23
networkx==1.11
oauth2client==4.0.0
packaging==16.8
pyasn1==0.1.9
pyasn1-modules==0.0.8
pyparsing==2.1.10
python-http-client==2.2.1
pytz==2016.10
PyYAML==3.11
requests==2.11.1
rsa==3.4.2
sendgrid==3.0.7
setuptools==34.0.2
six==1.10.0
uritemplate==3.0.0
urllib3==1.12
Werkzeug==0.11.15
